Services (#36)
* Reservations

* Default ReservationConfirmed

* remove pending

* Services

* auth and rbac

* services by business id

+++ api/reservation/models/reservationDto.go
@@ -9,7 +9,7 @@ import (
 type ReservationDto struct {
 	ID                uint                      `json:"id"`
 	AccountID         uint                      `json:"accountId"`
-	ServiceID         uint64                    `json:"serviceId"`
+	ServiceID         uint                      `json:"serviceId"`
 	DatePlaced        time.Time                 `json:"datePlaced"`
 	DateOfService     time.Time                 `json:"dateOfService"`
 	ReservationLength uint32                    `json:"reservationLength"`
